Major changes
- Dev price being calculated by number of followers and repositories.
- Contractor now can select number of work ours
- Minor bug fixes

+++ api/cart.php
@@ -1,5 +1,7 @@
 <?php
 
+include ('curl.php');
+
 if(!session_id()){
 	session_start();	
 }
@@ -12,16 +14,37 @@ if(isset($_GET["action"])){
 	$action = $_GET["action"];
 
 	if ($action == "add"){
-		$developer = $_POST;
-		add_to_cart($developer);
+		if(isset($_GET["username"]) && isset($_GET["price"])){
+			$username = $_GET["username"];
+			$price = $_GET["price"];
+			add_to_cart($username, $price);
+			update_order_total();
+		} else {
+			echo json_encode(array("message" => "Error: missing username parameter."));
+		}
 	}
 	else if ($action == "remove"){
 		if(isset($_GET["username"])){
 			remove_from_cart($_GET["username"]);
+			update_order_total();
+		} else {
+			echo json_encode(array("message" => "Error: missing username parameter."));
+		}
+
+	}
+	else if($action == "update"){
+		if(isset($_GET["username"]) && isset($_GET["hours"])){
+			$username = $_GET["username"];
+			$hours = $_GET["hours"];
+			update_user_total($username, $hours);
+			update_order_total();
+		} else {
+			echo json_encode(array("message" => "Error: missing username parameter."));
 		}
 	}
 	else if ($action == "clear"){
 		clear_cart();
+		update_order_total();
 	}
 	else if ($action == "get"){
 		get_cart();
@@ -29,7 +52,6 @@ if(isset($_GET["action"])){
 }
 
 function get_cart(){
-	
 	if(count($_SESSION["cart"]["products"]) == 0){
 		echo json_encode(array("message" => "Your cart is empty."));
 	} else {
@@ -37,20 +59,57 @@ function get_cart(){
 	}
 }
 
-function add_to_cart($developer){
+function add_to_cart($username, $price){
+	$already_in = false;
 	if(isset($_SESSION["cart"])){
-		$_SESSION["cart"]["products"][] = $developer;
-		echo json_encode(array("message" => "Developer added to cart."));
+		foreach ($_SESSION["cart"]["products"] as $i => $product) {
+			if($product->login == $username){
+				$already_in = true;
+			}
+		}
+		if($already_in){
+			echo json_encode(array("message" => "'$username' is already in your cart."));
+		} else {
+			$product = json_decode(curl_get_content("https://api.github.com/users/".$username));
+			$product->price = number_format($price, 2, ',', '');
+			$product->hours = 1;
+			$total_price = $price * $product->hours;
+			$product->total_price = number_format($total_price, 2, ',', '');
+			$_SESSION["cart"]["products"][] = $product;
+			echo json_encode(array("message" => "Developer added to cart."));
+		}
+		
+	}
+}
+
+function update_user_total($username, $hours){
+	foreach ($_SESSION["cart"]["products"] as $i => $product) {
+		if($product->login == $username){
+			$total_price = $product->price * $hours;
+			$product->total_price = number_format($total_price, 2, ',', ''); 
+			$product->hours = $hours;
+			echo json_encode(array("message" => "'$username' updated."));		
+		}
 	}
 }
 
 function remove_from_cart($username){
 	foreach ($_SESSION["cart"]["products"] as $i => $product) {
-		if($product["login"] == $username){
+		if($product->login == $username){
 			unset($_SESSION["cart"]["products"][$i]);
-			echo json_encode(array("message" => "'$username' removed from cart."));
+			echo json_encode(array("message" => "'$username' removed from cart."));		
 		}
 	}
+}
+
+function update_order_total(){
+
+		$order_total = 0;
+		foreach ($_SESSION["cart"]["products"] as $i => $product) {
+			$order_total = $order_total + $product->total_price;
+		}
+		$_SESSION["cart"]["order_total"] = number_format($order_total, 2, ',', '');
+
 }
 
 function clear_cart(){
